# backend/tevent/tevent_openapi.yml
openapi: 3.0.3
info:
  title: T-Event API
  description: |
    API для управления совместными событиями, расходами и взаиморасчетами.
    Система позволяет создавать события (группы), добавлять участников, 
    вести учет расходов и автоматически рассчитывать долги между участниками.
  version: 1.0.0
  contact:
    name: T-Event Team
    email: support@t-event.ru
servers:
  - url: https://api.t-event.ru/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - cookieAccessToken: []
...
paths:
...
  /events/{eventId}/expenses:
    get:
      tags:
        - Expenses
      summary: Получить список расходов события
      description: Возвращает данные события и список всех расходов
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные события и список расходов
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventHeader:
                    $ref: '#/components/schemas/EventHeaderDTO'
                  expenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExpenseDTO'

    post:
      tags:
        - Expenses
      summary: Добавить новый расход
      description: Создание нового расхода в событии
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Расход успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseDTO'
        '400':
          description: Ошибка валидации данных
  /expenses/{expenseId}/status:
    patch:
      tags:
        - Expenses
      summary: Изменить статус подтверждения расхода
      description: Подтверждение или оспаривание расхода участником
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [CONFIRMED, DISPUTED]
                comment:
                  type: string
                  description: Причина оспаривания (обязательно для DISPUTED)
      responses:
        '200':
          description: Статус успешно обновлен
        '404':
          description: Расход не найден
        '403':
          description: Недостаточно прав

  # ==================== ВЗАИМОРАСЧЕТЫ ====================
  /events/{eventId}/settlements:
    get:
      tags:
        - Settlements
      summary: Получить список взаиморасчетов
      description: Возвращает оптимизированный список переводов для закрытия долгов
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список расчетов
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventHeader:
                    $ref: '#/components/schemas/EventHeaderDTO'
                  settlements:
                    type: array
                    items:
                      $ref: '#/components/schemas/SettlementDTO'
  /settlements/{settlementId}/pay:
    post:
      tags:
        - Settlements
      summary: Отметить долг как оплаченный
      description: Должник отмечает, что перевел деньги получателю
      parameters:
        - name: settlementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Статус обновлен на WAITING_CONFIRMATION
        '404':
          description: Расчет не найден
        '400':
          description: Нельзя отметить как оплаченный (неверный статус)
  /settlements/{settlementId}/confirm:
    post:
      tags:
        - Settlements
      summary: Подтвердить получение средств
      description: Получатель подтверждает получение денег от должника
      parameters:
        - name: settlementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Долг успешно закрыт (статус COMPLETED)
        '404':
          description: Расчет не найден
        '400':
          description: Нельзя подтвердить (неверный статус)

  # ==================== ИСТОРИЯ ====================
...
components:
  securitySchemes:
    cookieAccessToken:
      type: apiKey
      in: cookie
      name: accessToken
      description: Access token для аутентификации
    cookieRefreshToken:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Refresh token для обновления access token
  schemas:
...
    EventHeaderDTO:
      type: object
      properties:
        title:
          type: string
          example: "Поездка в Сочи"
        startDate:
          type: string
          format: date
          example: "2026-06-15"
        endDate:
          type: string
          format: date
          example: "2026-06-20"
        status:
          type: string
          enum: [ACTIVE, COMPLETED, PLANNED]
          description: Статус события определяется автоматически на бэкенде на основе дат начала и окончания
        participantsCount:
          type: integer
          example: 5
        participantAvatars:
          type: array
          items:
            type: string
            format: uri
            description: Pre-signed URL из MinIO

    # ==================== УЧАСТНИКИ ====================
...
    ExpenseDTO:
      type: object
      properties:
        id:
          type: integer
          example: 101
        title:
          type: string
          example: "Бронирование отеля"
        category:
          type: string
          enum: [ACCOMMODATION, TRANSPORT, FOOD, ENTERTAINMENT, OTHER]
        approvalStatus:
          type: string
          enum: [CONFIRMED, PENDING, DISPUTED]
        payerName:
          type: string
          example: "Иван Петров"
        splitBetweenCount:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: double
          example: 45000.00
        perPersonAmount:
          type: number
          format: double
          example: 9000.00
        disputeInfo:
          type: object
          nullable: true
          properties:
            reason:
              type: string
              example: "Сумма указана неверно..."
            timestamp:
              type: string
              format: date-time
    CreateExpenseRequest:
      type: object
      required: [title, amount, category, participantIds]
      properties:
        title:
          type: string
          example: "Аренда машины"
        amount:
          type: number
          format: double
          example: 15000.00
        category:
          type: string
          enum: [ACCOMMODATION, TRANSPORT, FOOD, ENTERTAINMENT, OTHER]
          example: "TRANSPORT"
        participantIds:
          type: array
          items:
            type: integer
          description: ID участников, между которыми будет разделена сумма расхода
          example: [15, 23, 42]
          minItems: 1
        imageKey:
          type: string
          description: Ключ чека в MinIO
          example: "receipts/event_42/uuid.jpg"

    # ==================== ВЗАИМОРАСЧЕТЫ ====================
    SettlementDTO:
      type: object
      properties:
        id:
          type: integer
          example: 701
        fromUser:
          $ref: '#/components/schemas/UserShortDTO'
        toUser:
          $ref: '#/components/schemas/UserShortDTO'
        amount:
          type: number
          format: double
          example: 2000.00
        status:
          type: string
          enum: [PENDING, WAITING_CONFIRMATION, COMPLETED]
          description: |
            PENDING - долг висит (кнопка "Оплатить").
            WAITING_CONFIRMATION - должник нажал "Оплатить" (кнопка "Подтвердить" у получателя).
            COMPLETED - деньги получены (кнопка исчезает).
    UserShortDTO:
      type: object
      properties:
        id:
          type: integer
          example: 15
        fullName:
          type: string
          example: "Мария Сидорова"
        avatarUrl:
          type: string
          format: uri
        initials:
          type: string
          example: "М.С."

    # ==================== ИСТОРИЯ ====================
...